gollvm: add /usr/{lib,lib32} to toolchain paths
this updates the linux toolchain paths to include /usr/{lib,lib32}
to properly fix linking of crti.o on Arch Linux. This is where
Arch Linux puts certain files needed to link as can be seen on the
issue: https://github.com/golang/go/issues/36512 (this does not fix
that issue fully, but fixes one of the build errors).

I figured at first this could maybe be handled by the sysroot argument,
however that causes cmake to just fail to generate since ninja won't
be able to build a test binary, because other files are outside of
/usr/{lib,lib32}.

This seems like the easiest approach to fix this, without causing
errors for other distro's. That being said, I'm definetely not the
most knowledgable here, so if there's a better long term solution,
or a way to not hardcode `/usr/` in there like I have now, I'm happy
to change it, and test it on my system.

Updates golang/go#36512.

Change-Id: Ieea09f90c3e26ac79195cca6fe8fbc62a451b50a