Update patch set 1
Patch Set 1:

Thanks for sending this patch. Apologies for the delayed review (I was away on vacation, and I am still digging through large amounts of email). 

The code in gollvm that sets up library paths for the link is modeled after the corresponding code in clang; the clang code in question is extremely complicated and is riddled with special cases (I think mainly because of the many linux distros out there and the different ways in which they install/package gcc).

I dug through the clang source code again, and I don't see an obvious place where that code handles crt*.o files in /usr/{lib,lib64}, as they appear to be placed in arch linux. Clang does have a distro detector, and it does detect arch linux, but from what I can see the distro doesn't factor into the decision on how to set up the linker paths.

Cross compilation:  that's something we want to support eventually for llvm-goc, but it hasn't really been engineered yet.  With clang the model is that you use a single compiler binary for all targets (as opposed to GCC, where each target has a separate binary), but then you either build with --sysroot, or you insure that the target header/libraries are installed target-specific dirs (e.g. /usr/lib/gcc/<target>/<ver>/...).

I will do some more snooping, but it may be that your fix is the best way to go.

Patch-set: 1
Reviewer: Gerrit User 14020 <14020@62eb7196-b449-3ce5-99f1-c037f21e1705>