gollvm: add rudimentary linux distro detection support
Add support to the driver for detecting the linux distribution it is
running on. Different distributions install GCC headers and support
libraries in different places, complicating the process of selecting
the proper -L arguments to the linker. This patch will allow the
driver to look at the host's distro as part of the process of setting
up the library paths.

Updates golang/go#36512.

Change-Id: I90dfc8880096d63f1d36b97e068394d51936fe57