Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Hi, mind taking a look ? Thanks.

As I mentioned previously, for this to work correctly it needs to be handled with the helper function combineInputsWithEscapes (see the comment on that function for more info). This is necessary if you have an option that can be provided both directly to the driver (e.g. "-z foo") or in an escaped fashion to the linker (e.g. "-Wl,-z,foo").

Let me know if this is unclear. Thanks.

Patch-set: 4
Reviewer: Gerrit User 14020 <14020@62eb7196-b449-3ce5-99f1-c037f21e1705>