gollvm: implement getg function with inline assembly for arm64
The getg function is called frequently in runtime, and its performance
is very critical, so we turn the call to the getg function into a load
operation to inline it. In order to avoid the thread pointer caching
problem when thread switching happens between two function calls of this
function, we only inline the first one through GoSafeGetg pass. But this
implementation is not feasible on linux arm64, because it is based on the
assumption that the llvm backend's cse optimization of the thread pointer
only happens in a block range, but on linux arm64, this optimization occurs
in the entire function range.
Expanding GoSafeGetg pass to the entire function range is not a good
solution, because there will still be a large number of getg function
calls that are not inlined.
This CL simulates the implementation of getg in the c file through inline
assembly. This not only ensures its correctness, but also ensures that
each getg function call in Go files is inlined.
The disadvantage is that if no thread switching occurs between two getg
function calls, then theoretically the second getg function call can be
optimized, but not in this implementation. In this implementation, all
instruction sequences of getg function will be executed once it is called.

Updates golang/go#37295

Change-Id: If9e47b2afeb420a1d0316f2a82602b18bed82477