gollvm: Use aarch64linux emulation mode variants
When using the default ld linker on linux arm64, library libpthread.so can't be found.
Because currently we set emulation mode as aarch64elf, and its default search directories
only contain "/usr/aarch64-unknown-linux-gnu/lib", but libpthread.so is not there. We
can fix this issue by adding the correct path of this library to search list, but I'm
not sure if the path is fixed. As now we only support arm64 on linux, we simply replace
aarch64elf emulation mode with aarch64linux.

Change-Id: Ie779410d2e38f8d928a47cf31c1d68900eb19997