gollvm: set ld as the default linker
Currently gold is the default linker if -fuse-ld option isn't specified.
But gold on linux arm64 has a problem, which makes gollvm can not be used
for static compilation. On linux amd64, both ld and gold work well, and
gccgo set ld as default linker. For these reasons, this CL sets ld as the
default linker.

Change-Id: I79609efb015d94ab34e7b3f4b6fbd03bdf7ed757