gollvm: add GOLLVM_DEFAULT_LINKER cmake option
Currently gold is the default linker if -fuse-ld option isn't specified.
But gold on linux arm64 has a problem, which makes gollvm can not be used
for static compilation. This CL works around this issue by adding the option
--eh-frame-hdr to the linker when linking statically.

This CL also added a cmake option GOLLVM_DEFAULT_LINKER, which can be used
to configure the default linker.

Change-Id: I79609efb015d94ab34e7b3f4b6fbd03bdf7ed757