Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> I am ok with this in general. Might be worth adding a unit test.

I added a unit test for the dynamic link mode in the CL. I also wrote a test for the static link mode, but the return type of getg function is a temporary type, it's not a fixed value, so I didn't added it in this patch. Currently I didn't think of a good method to deal with this point.

Patch-set: 2