gollvm: fixes for unit test remastering machinery
Makes the unit test remastering feature a little more resilient with
respect to C++ compiler's interpretation of the line number for a
macro instantiation. If you have something like

   #define RAW_RESULT(x) #x
   #define PLINE(x) printf("%s %d\n", x, __LINE__)

and then instantiate with

  L10:   PLINE(R"RAW_RESULT(
  L11:     abc
  L12:     def
  L13:   )RAW_RESULT");

some C++ compilers will assign a value of 10 to the __LINE__
reference, and some will assign a value of 13. This patch adds
auto-detection for each of these two scenarios, and adjusts
accordingly.

Change-Id: Id9b29a6e13c6cab2dffe9432ed43200083daa29c