gollvm: rework handling of default linker
Revise the recipe for setting the default linker, setting the default
of "gold" in the cmake code instead of the tools code. As before with
no explicit setting of GOLLVM_DEFAULT_LINKER we'll target gold, but
GOLLVM_DEFAULT_LINKER can be set to "bfd" or "lld" (to select those
variants) or set to the empty string to just invoke "ld".

Change-Id: Ie8cca9244067b84d0841b01edb938b63ebc47053
Reviewed-on: https://go-review.googlesource.com/c/gollvm/+/239957
Reviewed-by: Cherry Zhang <cherryyz@google.com>