gollvm: rework handling of default linker
Tweak the recipe for selecting a default linker in cases where there
is no "-fuse-ld=" option on the command line. Previously we would
always default to ld.gold; now we default to "ld" unless
GOLLVM_DEFAULT_LINKER is set, in which case we use the variant
specific by that variable.

Change-Id: Ie8cca9244067b84d0841b01edb938b63ebc47053