gollvm: refactor unit test remastering code
Move some of the functionality that supports auto-remastering
of unit test results out of its current location in BackendCore
and make it part of TestUtils. This change is in preparation for
using it more widely in the unit tests (for example, in the
driver tests).

Change-Id: I9b2cb516483c3e8f3d5183367efb09d9dbcd1363
Reviewed-on: https://go-review.googlesource.com/c/gollvm/+/255048
Trust: Than McIntosh <thanm@google.com>
Reviewed-by: Cherry Zhang <cherryyz@google.com>