gollvm: add support for the LLVM integrated assembler
Add preliminary support for the integrated assembler (now enabled by
default, but selectable via -fintegrated-as/-fno-integrated-as).
Includes a set of new unit tests designed to verify that the driver is
doing the right thing in various scenarios.

Assemblers have many options (target-related options, options to
control DWARF/debug, etc); to support these options in the integrated
assembler the gollvm driver code has to look for them on the command
line (e.g. "-Wa,--fatal-warnings") and then set the appropriate flag
when it configures the LLVM assembly back end. The code to do this is
still largely unwritten at the moment, however.

The integrated assemble is on by default, but can be disable
using '-fno-integrated-as". The driver will also flip the default
(e.g. use the external assembler) if an unknown "-Wa,..."  is encountered
on the command line.

Change-Id: Id82d999187550d58b651dd51e03a7e9139aa50a5
Reviewed-on: https://go-review.googlesource.com/c/gollvm/+/255050
Reviewed-by: Cherry Zhang <cherryyz@google.com>
Trust: Than McIntosh <thanm@google.com>