gollvm: update libgo cmake rules to auto-generate linknames
This patch brings the gollvm cmake rules for libgo into sync with
recent gofrontend changes to the way //go:linkname directives are
emitted for various runtime packages (the corresponding gofrontend
change is in CL 265125). We now run an awk script over selected
packages to emit linknames for things mentioned in gen-sysinfo.go.

Change-Id: Iedf1c8b6d78033aa1e0b0009c6a74ef30aa99e72