gollvm: update recipe for recording sret/byval arg attributes
Update the recipe for attaching "sret" (structure return) and "by
value" attributes to function parameters. Previously it was enough to
just apply these attributes as tags; now we need to also capture the
type of the return or param as well.

Also fixes a buglet in the unit test remastering machinery (incorrect
env variable name).

Fixes golang/go#42843.

Change-Id: I3fceff4868f714d97621bdca07b91513c2ba94c8