gollvm: update recipe for recording SRET arg attribute
Update the recipe for attaching an attribute on to the function
parameter used to pass in memory used for the call return value (in
the case where we're passing things back in memory as opposed to in
registers). Previously it was enough to just apply the StructRet attr;
now we need to also capture the type of the return as well.

Also fixes a buglet in the unit test remastering machinery (incorrect
env variable name).

Fixes golang/go#42843.

Change-Id: I3fceff4868f714d97621bdca07b91513c2ba94c8